Bypass pyne exception reporting if run inside a debugger.
When running inside of a debug environment, pyne does a
very unuseful thing of going to where the TestFailureException is
thrown. If pyne was able to determine that it is being run inside a
debug environment (via pyne_pdb.is_running_inside_debugger()), then it
does not perform a try/except when running code, allowing the debugger
to jump to the actual exception location.