Add ability to specify method_name for stub/Spy
Sometimes the __name__ of a function is not the name that it can
referenced by on its parent object. This happens for example when
decorators are used.

class MyClass:
    def my_method(self):
        pass

pyne assumes my_class.my_method has name "my_method".

This assumption is violated if a decorator is placed on "my_method"

class MyClass:
    @some_decorator
    def my_method(self):
        pass

my_class.my_method will now have a name related to the decorator, and
will not be able to correctly replace the function with a spy.