Add an AttachedSpy class that works with any method on any object
There are several points of fragility with the current spy class:

1) It its reliance on introspection to determine the method name
2) The assumption that the method's name matches its name on its object
3) The assumption that swapping a function for a Spy object always works

The AttachedSpy is a stepping stone to having a spy that "just works'.

It seperates concerns from how to get the method name and parent object
from the actually stubbing behavior, and makes sure to stub out the
original method with another method instead of an object.

It addresses fragility as follows
1) It its reliance on introspection to determine the method name
This logic is not done. One can use the AttachedSpy by manually
specifying the parent object and the method name. This is useful in the
cases where the normal spy introspection fails. (A method_name option was previously
added to the spy to help in this case).

In the future, much of the introspection logic can be done in the spy
helper function, and then the AttachedSpy created with its results.

One way that introspection fails is with decorated functions, who's name
is that of the function wrapper (sometimes a lambda).

2) The assumption that the method's name matches its name on its object
This also happens with decorators, but may happen in other cases two.
In this case, the Spy is placed at an unexpected location and will not
be invoked it test cases. When an object is stubbed soley for the
purpose of suppressing the original function behavior, the original
function will be invoked without warning.

This is by-passed by specifying the name of the method's attachment
point on the parent class, with the same rational as 1).

3) The assumption that swapping a function for a Spy object always works
In general it is okay to replace a function for a Spy. However, there is
a valid use case to stub the an unbound class method when one would like
to track the behavior of a bound class method, but the instance of which
is unobtainable in the test code. One would stub SomeClass.some_method,
   and then indirectly spy on some_instance.some_method's calls. When
   some_method is replaced by a spy, the normal binding behaviour of
   python class methods breaks, resulting in call failures if the
   original method is to be called as "self" is unavailable to the
   unbound spy.

  The implementation of this is to replace the target method with a
  bound "spy method" instead of the spy itself. The downside is that
  spy attributes can't be accessed directly from this spy method.
  Although the spy can be accessed by calling Spy.get_spy(some_method).
  This is how expectations are still possible using the attached
  function, such as
```
 expect(some_instance.some_method).to_be_called()
```

Code such as
```
 some_instance.some_method.last_call
```
isn't valid with an attached spy.
One must do:
```
Spy.get_spy(some_instance.some_method).last_call
```
Internally, this calls the spy function with a special kwarg that
triggers it to return self.