Suppress trackback print of TestFailureException
When running a single pyne test, there is nothing to catch the
TestFailureException as happens in the cli tool. Instead, it gets
printed without formatting and makes the test results hard to read.
We can make it look cleaner by only printing the exception value.
This is done by overriding sys.excepthook and adding special printing
logic for TestFailureException.