Add MegaStub to replace unintuitive group_stubs
`group_stubs` was previously added to manage a collection of stubs as a group.
It gives the impression that it can be used to manage the stubs as a
group, but in practice it only:
1) Applies all stubs on creation
2) Will remove all stubs on exit of a with statement

Instead of fixing this behavior, and breaking tests that rely on it, it
is "deprecated", and a new `mega_stub` method creates an object that
manages the collection of stubs.
1) Doesn't change the state of the stubs on creation
2) Has `stub` and `unstub` methods that applies to the group
3) Has with-statement support that applies to the group

Stub / Unstab functions were added to Spy so that it can be used as part
of a `MegaStub`

`group_stubs` is marked as deprecated and yields a warning